requirements-aux.txt
flask
flask-restx
requests
pyjwt==2.1.0
aiohttp==3.7.4.post0
asyncio==3.4.3
Werkzeug==2.3.7
python-dotenv